Extract to subdir w/ a+r perms and extract menu
This is a working plugin that makes the extracted files readable by all
users and extracted into a subfolder `./extract/`.

Also adds ability to right click torrent and select Extract from context
menu. Theres a lot of stuff that could be cleaned up but for now it's
working for my needs.